# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenPrompt API",
    "description": "REST API for prompt optimization, evaluation, and benchmarking.",
    "version": "0.2.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/lint": {
      "post": {
        "summary": "Lint Endpoint",
        "operationId": "lint_endpoint_lint_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/optimize": {
      "post": {
        "summary": "Optimize Endpoint",
        "operationId": "optimize_endpoint_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/evaluate": {
      "post": {
        "summary": "Evaluate Endpoint",
        "operationId": "evaluate_endpoint_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/benchmark": {
      "post": {
        "summary": "Benchmark Endpoint",
        "operationId": "benchmark_endpoint_benchmark_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/compress": {
      "post": {
        "summary": "Compress Endpoint",
        "operationId": "compress_endpoint_compress_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/multi-model/optimize": {
      "post": {
        "summary": "Multi Model Endpoint",
        "operationId": "multi_model_endpoint_multi_model_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultiModelOptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiModelOptimizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cost/recommend": {
      "post": {
        "summary": "Cost Recommend Endpoint",
        "operationId": "cost_recommend_endpoint_cost_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostRecommendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostRecommendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BenchmarkRequest": {
        "properties": {
          "prompts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prompts"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "mock"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "mock-model"
          }
        },
        "type": "object",
        "required": [
          "prompts"
        ],
        "title": "BenchmarkRequest"
      },
      "BenchmarkResponse": {
        "properties": {
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "entries": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Entries"
          },
          "markdown": {
            "type": "string",
            "title": "Markdown"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "entries",
          "markdown"
        ],
        "title": "BenchmarkResponse"
      },
      "CompressRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "mock"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "mock-model"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "CompressRequest"
      },
      "CostRecommendRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy",
            "default": "rewrite"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "mock"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "mock-model"
          },
          "min_quality": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Quality"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "CostRecommendRequest"
      },
      "CostRecommendResponse": {
        "properties": {
          "recommended": {
            "additionalProperties": true,
            "type": "object",
            "title": "Recommended"
          },
          "pareto_frontier": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Pareto Frontier"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "quality_per_dollar": {
            "type": "number",
            "title": "Quality Per Dollar"
          }
        },
        "type": "object",
        "required": [
          "recommended",
          "pareto_frontier",
          "reason",
          "quality_per_dollar"
        ],
        "title": "CostRecommendResponse"
      },
      "EvaluateRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "mock"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "mock-model"
          },
          "tests": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tests"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "tests"
        ],
        "title": "EvaluateRequest"
      },
      "EvaluateResponse": {
        "properties": {
          "accuracy": {
            "type": "number",
            "title": "Accuracy"
          },
          "pass_rate": {
            "type": "number",
            "title": "Pass Rate"
          },
          "prompt_tokens": {
            "type": "integer",
            "title": "Prompt Tokens"
          },
          "total_cost_usd": {
            "type": "number",
            "title": "Total Cost Usd"
          },
          "total_latency_ms": {
            "type": "number",
            "title": "Total Latency Ms"
          },
          "judge_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Score"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "accuracy",
          "pass_rate",
          "prompt_tokens",
          "total_cost_usd",
          "total_latency_ms",
          "results"
        ],
        "title": "EvaluateResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "version"
        ],
        "title": "HealthResponse"
      },
      "LintIssueResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "severity"
        ],
        "title": "LintIssueResponse"
      },
      "LintRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "LintRequest"
      },
      "LintResponse": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/LintIssueResponse"
            },
            "type": "array",
            "title": "Issues"
          },
          "categories": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Categories"
          }
        },
        "type": "object",
        "required": [
          "score",
          "issues"
        ],
        "title": "LintResponse"
      },
      "ModelSpecRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model"
        ],
        "title": "ModelSpecRequest"
      },
      "MultiModelOptimizeRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "models": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ModelSpecRequest"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "array",
            "title": "Models"
          },
          "strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy",
            "default": "rewrite"
          },
          "tests": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tests"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "models"
        ],
        "title": "MultiModelOptimizeRequest"
      },
      "MultiModelOptimizeResponse": {
        "properties": {
          "markdown_table": {
            "type": "string",
            "title": "Markdown Table"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Rows"
          },
          "best_quality_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Quality Model"
          },
          "lowest_cost_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lowest Cost Model"
          }
        },
        "type": "object",
        "required": [
          "markdown_table",
          "rows"
        ],
        "title": "MultiModelOptimizeResponse"
      },
      "OptimizeRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy",
            "default": "hybrid"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "mock"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "mock-model"
          },
          "tests": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tests"
          },
          "objective": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "constraints": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraints"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "OptimizeRequest"
      },
      "OptimizeResponse": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "original_score": {
            "type": "number",
            "title": "Original Score"
          },
          "optimized_score": {
            "type": "number",
            "title": "Optimized Score"
          },
          "score_delta": {
            "type": "number",
            "title": "Score Delta"
          },
          "original_tokens": {
            "type": "integer",
            "title": "Original Tokens"
          },
          "optimized_tokens": {
            "type": "integer",
            "title": "Optimized Tokens"
          },
          "token_delta_pct": {
            "type": "number",
            "title": "Token Delta Pct"
          },
          "original_cost_usd": {
            "type": "number",
            "title": "Original Cost Usd"
          },
          "optimized_cost_usd": {
            "type": "number",
            "title": "Optimized Cost Usd"
          },
          "cost_delta_pct": {
            "type": "number",
            "title": "Cost Delta Pct"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "report_lines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Report Lines"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "original_score",
          "optimized_score",
          "score_delta",
          "original_tokens",
          "optimized_tokens",
          "token_delta_pct",
          "original_cost_usd",
          "optimized_cost_usd",
          "cost_delta_pct",
          "strategy"
        ],
        "title": "OptimizeResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}